Refactor promises to use await

+++ src/controller/story/deleteStoryController.ts
@@ -2,9 +2,12 @@ import { Response } from "express";
 import { storyDao } from "../../dao/storyDao.js";
 import { AuthenticatedRequest } from "../../middleware/authenticate.js";
 
-export function deleteStoryController(req: AuthenticatedRequest, res: Response) {
-    storyDao.deleteStory(req.username, req.params.storyId)
-        .then(() => {
-            res.sendStatus(200);
-        });
+export async function deleteStoryController(req: AuthenticatedRequest, res: Response) {
+    if (!req.username) {
+        res.sendStatus(409);
+        return;
+    }
+
+    await storyDao.deleteStory(req.username, req.params.storyId);
+    res.sendStatus(200);
 }
